Reference

Content

Index

VegSci.SyntopicTable

VegSci.binary_dissimilarity

VegSci.compose_syntopic_table_object

VegSci.correspondence_analysis

VegSci.generate_test_array

VegSci.indval_fidelity

VegSci.jaccard_coefficient

VegSci.nzfunc

Public Interface

VegSci.SyntopicTable — Type

SyntopicTable

Fields

name – The names of the association, of class String

releve_n – The number of releves used to compose the syntopic table object, of class Int64

releve_ids – The ids of the releves used in the construction of the syntopic table object, usually the row names of the releve by species matrix, of class Vector{String}

species_n – The number of species in the syntopic table, of class Int64

species_names – The names of species in the syntopic table, of class Vector{String}

abundance_units – The abundance units, of class String

relative_frequency – The relative frequency of occurrence of each species across all releves, of class Vector{Float64}

absolute_frequency – Vector{Float64}

minimum_abundance – Vector{Float64}

mean_abundance – Vector{Float64}

median_abundance – Vector{Float64}

maximum_abundance – Vector{Float64}

fidelity – Vector{Float64}

fidelity_p – Vector{String}

fidelity_n – Vector{String}

VegSci.compose_syntopic_table_object — Function

composesyntopictable_object(name::String, x::NamedArray)::SyntopicTable

Compose a single syntopic table object from a releve by species matrix

Input

name – The name given to the association represented in the sytopic table, of class String

x – A releve by species matrix, of class NamedArrays::NamedMatrix

cover_abundance_scale – The units of the matrix values, of class String

Output

A object of class VegSci.SyntopicTable. See ...

Notes

Algorithm

References

VegSci.correspondence_analysis — Method

correspondence_analysis(N::NamedMatrix)::NamedMatrix

Perform a correspondence analysis following the computational algorithm outlined in appendix A of Greenacre (2017).

Input

N – A releve by species matrix of the class NamedArrays::NamedMatrix

Output

Notes

Algorithm

References

Greenacre, M., 2017. Correspondence Analysis in Practice, Third Edition. CRC Press.

VegSci.generate_test_array — Method

greet_your_package_name(name)

Create a rown by coln NamedArray object containing random values.

...

Arguments

rown::Int64: The number of rows in the array.

coln::Int64: The number of columns in the array.

meancoloccs::Int64: The mean number of non-zero elements in the array.

rowprefix::String: The prefix to the row number. "Releve" by default.

colprefix::String: The prefix to the column number. "Species" by default.

rowdim::String: The row dimension name. "Releve" by default.

coldim::String: The column dimension name. "Species" by default.

...

Examples

julia>generate_test_array(rown = 10, coln = 10, meancoloccs = 5, rowprefix = "Releve", colprefix = "Species")

VegSci.indval_fidelity — Method

indval_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}})::NamedMatrix

Calculate the Dufrêne-Legendre Indicator Value Index (IndVal) fidelity values for a given releve by species matrix and set of clusters.

Input

x – A releve by species matrix of the class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster of the class Dict{String, Vector{String}}

Output

A cluster by species NamedMatrix containing the fidelity values.

Notes

Algorithm

This function implements the Dufrêne-Legendre Indicator Value Index as described in Dufrêne & Legendre (1997)

References

Dufrêne, M. & Legendre, P. 1997. Species assemblages and indicator species: the need for a flexible asymmetrical approach. Ecol. Monogr. 67: 345-366.

VegSci.jaccard_coefficient — Method

VegSci.nzfunc — Method

Draws heavily from the function outlines here: https://discourse.julialang.org/t/nanmean-options/4994/17

Internal functions

VegSci.binary_dissimilarity — Method

binary_dissimilarity(x::Union{Matrix, NamedMatrix}, eq::AbstractString)

Create a personalised greeting for VegSci using a name.

...

Arguments

x::Union{Matrix, NamedMatrix}: A releve by species matrix of the class Matrix or NamedArrays.NamedMatrix

eq::AbstractString: A representing an equation using the terms a, b, and c. The Jaccard similarity by default. "(a ./ (a .+ b .+ c)) + I".

...

Examples

x = generatetestarray(rown = 10, coln = 10)
julia>binarysimilarity(x, "(a ./ (a .+ b .+ c)) + I")