Reference

Content

Index

VegSci.SyntopicTable

VegSci.G_fidelity

VegSci.binary_dissimilarity

VegSci.check_relSpp_array_format

VegSci.chisq_fidelity

VegSci.compose_syntopic_table_object

VegSci.correspondence_analysis

VegSci.extract_indicators

VegSci.generate_test_array

VegSci.indval_fidelity

VegSci.jaccard_coefficient

VegSci.nzfunc

VegSci.phi_fidelity

VegSci.u_fidelity

Public Interface

VegSci.SyntopicTable — Type

SyntopicTable

Fields

name – The name of the association, of class String

code – The code for the association, of class String

releve_n – The number of releves used to compose the syntopic table object, of class Int64

releve_ids – The ids of the releves used in the construction of the syntopic table object, usually the row names of the releve by species matrix, of class Vector{String}

species_n – The number of species in the syntopic table, of class Int64

species_names – The names of species in the syntopic table, of class Vector{String}

minimum_species – The minimum species richness within the releves, of class Int64

mean_species – The mean species richness across the releves, of class Float64

maximum_species – The maximum species richness within the releves, of class Int64

abundance_units – The abundance units, of class String

relative_frequency – The relative frequency of occurrence of each species across all releves, of class Vector{Float64}

absolute_frequency – Vector{Float64}

minimum_abundance – Vector{Float64}

mean_abundance – Vector{Float64}

median_abundance – Vector{Float64}

maximum_abundance – Vector{Float64}

fidelity – Vector{Float64}

fidelity_p – Vector{String}

fidelity_n – Vector{String}

VegSci.G_fidelity — Function

G_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}}, method::String = "G")::NamedMatrix

Calculate the G statistic fidelity values for a given releve by species matrix and set of clusters.

Input

x – A releve by species matrix of the class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster of the class Dict{String, Vector{String}}

method – The G statistic fidelity method to use one of: "G" or "G_adj", of class String

Output

A cluster by species NamedMatrix containing the fidelity values.

Notes

Algorithm

This function implements the G statistic fidelity measure as described in Chytrý et al (2002) and Sokal and Rohlf (1995).

References

Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x Sokal, R.R., Rohlf, F.J., 1995. Biometry. W. H. Freeman.

VegSci.check_relSpp_array_format — Method

check_relSpp_array_format(x::NamedArray; lower_bound::Float64 = 0.0, upper_bound::Float64 = 1.0)

Check whether the format of a Releve by Species matrix is correct. Specifically whether the following conditions are met: - All column names are strings - All row names are strings - Cover-Abundance values are above or equal to a lower bound - Cover-Abundance values are below or equal to a upper bound - There are no NaN Cover-Abundance values - There are no missing Cover-Abundance values

...

Arguments

x::NamedArray: A Releve by Species matrix

lower_bound::Float64: The minimum allowable cover-abundance value.

upper_bound::Float64: The maximum allowable cover-abundance value.

...

Examples

julia> x = generate_test_array(rown = 10, coln = 10, meancoloccs = 5, rowprefix = "Releve", colprefix = "Species")
julia> x = check_relSpp_array_format(x)

VegSci.chisq_fidelity — Function

chisq_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}}, method::String = "chisq")::NamedMatrix

Calculate the Chi-squared fidelity values for a given releve by species matrix and set of clusters.

Input

x – A releve by species matrix of the class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster of the class Dict{String, Vector{String}}

method – The Chi-squared fidelity method to use one of: "chisq" or "chisq_adj", of class String

Output

A cluster by species NamedMatrix containing the fidelity values.

Notes

Algorithm

This function implements the Chi-squared fidelity measure as described in Chytrý et al (2002) and Sokal and Rohlf (1995).

References

Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x Sokal, R.R., Rohlf, F.J., 1995. Biometry. W. H. Freeman.

VegSci.compose_syntopic_table_object — Function

composesyntopictable_object(name::String, x::NamedArray)::SyntopicTable

Compose a single syntopic table object from a releve by species matrix

Input

name – The name given to the association represented in the sytopic table, of class String

x – A releve by species matrix, of class NamedArrays::NamedMatrix

cover_abundance_scale – The units of the matrix values, of class String

Output

A object of class VegSci.SyntopicTable.

Notes

Algorithm

References

VegSci.correspondence_analysis — Method

correspondence_analysis(N::NamedMatrix)::NamedMatrix

Perform a correspondence analysis following the computational algorithm outlined in appendix A of Greenacre (2017).

Input

N – A releve by species matrix of the class NamedArrays::NamedMatrix

Output

Notes

Algorithm

References

Greenacre, M., 2017. Correspondence Analysis in Practice, Third Edition. CRC Press.

VegSci.extract_indicators — Method

extract_indicators(fidelity_values::NamedMatrix; p_cut::Float64, n_cut::Union{Float64, Nothing} = nothing)

Extract the positive and negative indicator species from a matrix of fidelity values using a positive and optionally a negative cut value. Positive indicators are denoted by a value of 1.0 and negative indicators are denoted by a value of -1.0.

Input

x – A releve by species matrix, of class NamedArrays::NamedMatrix

p_cut – A postive number used to filter positive indicator species, of class Float64

n_cut – A number used to filter negative indicator species, of class Float64

Output

A cluster by species NamedMatrix containing the positive (1.0) and negative (-1.0) indicator species.

Notes

VegSci.generate_test_array — Method

generate_test_array(;rown::Int64, coln::Int64,
                    meancoloccs::Int64,
                    rowprefix::String = "Releve", colprefix::String = "Species",
                    rowdim::String = "Releve", coldim::String = "Species")

Create a rown by coln NamedArray object containing random values.

...

Arguments

rown::Int64: The number of rows in the array.

coln::Int64: The number of columns in the array.

meancoloccs::Int64: The mean number of non-zero elements in the array.

rowprefix::String: The prefix to the row number. "Releve" by default.

colprefix::String: The prefix to the column number. "Species" by default.

rowdim::String: The row dimension name. "Releve" by default.

coldim::String: The column dimension name. "Species" by default.

...

Examples

julia>generate_test_array(rown = 10, coln = 10, meancoloccs = 5, rowprefix = "Releve", colprefix = "Species")

VegSci.indval_fidelity — Method

indval_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}})::NamedMatrix

Calculate the Dufrêne-Legendre Indicator Value Index (IndVal) fidelity values for a given releve by species matrix and set of clusters.

Input

x – A releve by species matrix of the class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster of the class Dict{String, Vector{String}}

Output

A cluster by species NamedMatrix containing the fidelity values.

Notes

Algorithm

This function implements the Dufrêne-Legendre Indicator Value Index as described in Dufrêne & Legendre (1997) and Chytrý et al (2002).

References

Dufrêne, M. & Legendre, P. 1997. Species assemblages and indicator species: the need for a flexible asymmetrical approach. Ecol. Monogr. 67: 345-366. Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x

VegSci.jaccard_coefficient — Method

VegSci.nzfunc — Method

Draws heavily from the function outlines here: https://discourse.julialang.org/t/nanmean-options/4994/17

VegSci.phi_fidelity — Method

phi_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}})::NamedMatrix

Calculate the Φ fidelity values for a given releve by species matrix and set of clusters.

Input

x – A releve by species matrix of the class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster of the class Dict{String, Vector{String}}

Output

A cluster by species NamedMatrix containing the fidelity values.

Notes

Algorithm

This function implements the Φ fidelity measure as described in Chytrý et al (2002) and Sokal and Rohlf (1995).

References

Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x Sokal, R.R., Rohlf, F.J., 1995. Biometry. W. H. Freeman.

VegSci.u_fidelity — Function

u_fidelity(x::NamedMatrix, clusters::Dict{String, Vector{String}}, method::String)::NamedMatrix

Calculate a Bruelheide u fidelity measure.

Input

x – A releve by species matrix, of class NamedArrays::NamedMatrix

clusters – A dictionary containing the names of the clusters and associated lists of row names belonging to that cluster, of class Dict{String, Vector{String}}

method – The u fidelity method to use one of: "uhyp", "uhypadj", "ubinA", "ubinAadj", "ubinB", "ubinB_adj", of class String

Output

A cluster by species NamedMatrix containing the selected u fidelity values.

Notes

Algorithm

This function implements the Bruelheide fidelity measure of u as outlined in Bruelheide (2000) and Chytrý et al (2002).

References

Bruelheide, H., 2000. A new measure of fidelity and its application to defining species groups. Journal of Vegetation Science 11, 167–178. https://doi.org/10.2307/3236796 Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x

Internal functions

VegSci.binary_dissimilarity — Method

binary_dissimilarity(x::Union{Matrix, NamedMatrix}, eq::AbstractString)

Create a personalised greeting for VegSci using a name.

...

Arguments

x::Union{Matrix, NamedMatrix}: A releve by species matrix of the class Matrix or NamedArrays.NamedMatrix

eq::AbstractString: A representing an equation using the terms a, b, and c. The Jaccard similarity by default. "(a ./ (a .+ b .+ c)) + I".

...

Examples

x = generatetestarray(rown = 10, coln = 10)
julia>binarysimilarity(x, "(a ./ (a .+ b .+ c)) + I")